Move ShouldBeSimilarTo
