+++ src/GraphQL.Tests/Types/InputObjectGraphTypeTests.cs
@@ -1,6 +1,5 @@
 #nullable enable
 
-using GraphQL.Tests.Subscription;
 using GraphQL.Types;
 
 namespace GraphQL.Tests.Types;
